Move rough draft out of test directory
